# tools/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Investment Tracker API
  version: 0.1.0
  summary: AI-native portfolio tracker REST API
  description: |
    Read-only aggregator for investment portfolios with an AI-first UX.

    ## Design principles

    - **Spec-first.** This file is the source of truth. Go (huma v2), TypeScript
      (openapi-typescript / openapi-fetch) and Swift (swift-openapi-generator)
      artefacts are generated from this document.
    - **All paths under `/v1/`.** Breaking changes go to `/v2/`.
    - **Cursor pagination** everywhere a list is returned.
    - **Standard error envelope** with a stable `code` and a `request_id`.
    - **Rate limits** are communicated via `X-RateLimit-*` headers on every
      response; `429` is returned when exceeded.
    - **Idempotency.** All mutating endpoints (`POST`, `PATCH`, `DELETE`) accept
      an optional `Idempotency-Key` header. The server stores the response for
      24 hours and replays it for a repeated key.
    - **Money is a string.** Decimal strings (e.g. `"123.4567890000"`) preserve
      precision; JSON numbers would not.
    - **Multi-currency.** Values are always returned in both `base` (USD today)
      and `display` (user preference, converted on the fly).
    - **SSE streaming.** `POST /v1/ai/chat/stream` uses `text/event-stream`.
      A non-stream `POST /v1/ai/chat` counterpart exists for tests and
      retrospectives.
  contact:
    name: Investment Tracker Team
    email: dev@investmenttracker.example
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /me/export:
    get:
      tags: [Users]
      operationId: exportMe
      summary: GDPR data export
      description: |
        Returns a JSON bundle of every piece of personal data held for the
        current user. Suitable for download by the user via the app.
      responses:
        '200':
          description: Data export bundle
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserExportBundle' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /market/quote:
    get:
      tags: [Market]
      operationId: getMarketQuote
      summary: Get a single quote
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
        - name: asset_type
          in: query
          required: true
          schema: { $ref: '#/components/schemas/AssetType' }
      responses:
        '200':
          description: Quote
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketQuote' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /exports:
    post:
      tags: [Users]
      operationId: createExport
      summary: Queue a CSV export job
      description: |
        Async-only (see design brief §15.5). Returns `202` with a job id;
        clients poll `GET /exports/{id}` until `status=done` and then
        download `result_url`. Tier-gated: Plus+.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportCreateRequest' }
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportJob' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: CSV export requires Plus or higher
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /exports/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Users]
      operationId: getExport
      summary: Poll an export job
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────── Analytics & Tax (Pro) ─────────────────
...
  /portfolio/tax/export:
    get:
      tags: [Portfolio]
      operationId: exportPortfolioTax
      summary: Download a tax report as PDF or CSV
      parameters:
        - $ref: '#/components/parameters/TaxJurisdiction'
        - $ref: '#/components/parameters/TaxYear'
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [pdf, csv]
      responses:
        '200':
          description: |
            Returns the rendered file directly for PDF/CSV; clients should
            trigger a download rather than parse the body.
          content:
            application/pdf:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Tax export requires Pro
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ─────────────────────────── Glossary ──────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer JWT issued by Clerk. Validated against Clerk JWKS on every
        request. `sub` claim is mapped to `users.clerk_user_id`.

  # ─────────────────────── Reusable parameters ───────────────────
  parameters:
...
    TaxJurisdiction:
      name: jurisdiction
      in: query
      required: true
      description: ISO-3166 alpha-2 country code for the tax regime.
      schema:
        type: string
        enum: [US, DE, FR, UK, ES, NL]
    TaxYear:
      name: year
      in: query
      required: true
      schema:
        type: integer
        minimum: 2000
        maximum: 2100
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Opaque client-generated key (UUID v4 recommended, max 255 chars).
        The server stores the response for 24 hours and replays it for a
        repeated key on the same authenticated user + same endpoint. Keys
        are ignored on `GET` requests.
      schema: { type: string, maxLength: 255 }

  # ─────────────────────── Reusable headers ──────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Resource does not exist or is not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    Money:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,10})?$'
      description: |
        Decimal money string, up to 10 fractional digits. Matches the
        `NUMERIC(30, 10)` precision used in the database.
      examples: ['123.4567890000', '-5.0', '0']
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      description: ISO 4217 currency code.
      examples: ['USD', 'EUR', 'GBP', 'JPY', 'BTC', 'USDT']
    AssetType:
      type: string
      enum: [stock, etf, crypto]
    TransactionType:
      type: string
      enum: [buy, sell, dividend, split, transfer_in, transfer_out, fee]
    ConnectionType:
      type: string
      enum: [api, aggregator, import, manual]
    AccountType:
      type: string
      enum: [broker, crypto, manual]
    SyncState:
      type: string
      enum: [pending, syncing, ok, error, paused]
      description: |
        - `pending` — created, not yet synced.
        - `syncing` — a sync is in-flight.
        - `ok` — last sync succeeded.
        - `error` — last sync failed; user action needed (§18.2).
        - `paused` — user-initiated pause; cron skips this account until
          `POST /accounts/{id}/resume`.
    SubscriptionTier:
      type: string
      enum: [free, plus, pro]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: |
                Stable machine-readable error code.

                - `TIER_LIMIT_EXCEEDED` — quota exhausted on the current tier
                  (e.g. 5/5 AI messages today). UI shows usage-reset messaging.
                - `FEATURE_LOCKED` — tier does not include this feature at all
                  (e.g. Free/Plus calling `/portfolio/analytics`). UI shows
                  the upgrade/blur-preview treatment from design brief §13.5.
              enum:
                - VALIDATION_ERROR
                - INVALID_REQUEST
                - UNAUTHENTICATED
                - FORBIDDEN
                - TIER_LIMIT_EXCEEDED
                - FEATURE_LOCKED
                - NOT_FOUND
                - CONFLICT
                - DUPLICATE_TRANSACTION
                - IDEMPOTENCY_KEY_CONFLICT
                - IDEMPOTENCY_IN_PROGRESS
                - JURISDICTION_NOT_SUPPORTED
                - QUOTE_NOT_AVAILABLE
                - RATE_LIMIT_EXCEEDED
                - EXTERNAL_SERVICE_ERROR
                - NOT_IMPLEMENTED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable message. Do not parse this.
            details:
              type: object
              additionalProperties: true
              description: Structured details for the error (e.g. failing field).
            request_id:
              type: string
              description: Correlation id; include this in bug reports.
...
    User:
      type: object
      required:
        - id
        - clerk_user_id
        - email
        - display_currency
        - locale
        - subscription_tier
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        clerk_user_id: { type: string }
        email: { type: string, format: email }
        display_currency: { $ref: '#/components/schemas/Currency' }
        locale: { type: string, examples: ['en', 'en-US', 'ru'] }
        subscription_tier: { $ref: '#/components/schemas/SubscriptionTier' }
        stripe_customer_id:
          type: [string, "null"]
        deletion_scheduled_at:
          type: [string, "null"]
          format: date-time
          description: |
            Set when the user initiated account deletion (see §17.4).
            The hard-delete worker runs after a 7-day grace period; the user
            can call `POST /me/undo-deletion` within that window to cancel.
            `null` means the account is active / not pending deletion.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    UserExportBundle:
      type: object
      required: [generated_at, user, accounts, transactions]
      additionalProperties: true
      properties:
        generated_at: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
        accounts:
          type: array
          items: { $ref: '#/components/schemas/Account' }
        transactions:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }

    # ── Account ───────────────────────────────────────────────
    Account:
      type: object
      required:
        - id
        - broker_name
        - display_name
        - account_type
        - connection_type
        - base_currency
        - sync_status
        - is_included_in_portfolio
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        broker_name:
          type: string
          description: |
            Canonical broker/exchange identifier. Free-text to allow new
            integrations without a spec bump; known values include
            `interactive_brokers`, `schwab`, `fidelity`, `binance`,
            `coinbase`, `kraken`, `manual`.
          examples: [interactive_brokers, binance]
        display_name:
          type: string
          maxLength: 120
          description: User-editable label.
        account_type: { $ref: '#/components/schemas/AccountType' }
        connection_type: { $ref: '#/components/schemas/ConnectionType' }
        external_account_id:
          type: [string, "null"]
        base_currency: { $ref: '#/components/schemas/Currency' }
        last_synced_at:
          type: [string, "null"]
          format: date-time
        sync_status: { $ref: '#/components/schemas/SyncState' }
        sync_error:
          type: [string, "null"]
        is_included_in_portfolio:
          type: boolean
          description: |
            When false, the account is hidden from portfolio aggregations
            (`/portfolio`, `/positions`, snapshots) but still listed in
            `/settings/accounts` (§18.1). Toggled via `PATCH /accounts/{id}`.
        deleted_at:
          type: [string, "null"]
          format: date-time
          description: Present on tombstoned rows (usually omitted from live list queries).
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Transaction:
      type: object
      required:
        - id
        - account_id
        - symbol
        - asset_type
        - transaction_type
        - quantity
        - currency
        - executed_at
        - source
        - created_at
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        symbol: { type: string, examples: [AAPL, BTC, VWCE] }
        asset_type: { $ref: '#/components/schemas/AssetType' }
        transaction_type: { $ref: '#/components/schemas/TransactionType' }
        quantity: { $ref: '#/components/schemas/Money' }
        price:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: "null"
          description: Null for `split` events.
        currency: { $ref: '#/components/schemas/Currency' }
        fee: { $ref: '#/components/schemas/Money' }
        executed_at: { type: string, format: date-time }
        source:
          type: string
          enum: [manual, api, aggregator, import]
        source_details:
          type: [object, "null"]
          additionalProperties: true
          description: |
            Loose provider-specific payload (e.g. SnapTrade `trade_id`, CSV
            row number). Not intended to be consumed by clients.
        fingerprint:
          type: string
          description: Opaque dedup hash.
        notes:
          type: [string, "null"]
          maxLength: 1000
        created_at: { type: string, format: date-time }
...
    MarketQuote:
      type: object
      required: [symbol, asset_type, currency, price, as_of, source]
      properties:
        symbol: { type: string }
        asset_type: { $ref: '#/components/schemas/AssetType' }
        currency: { $ref: '#/components/schemas/Currency' }
        price: { $ref: '#/components/schemas/Money' }
        previous_close:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: "null"
        change_percent:
          type: [number, "null"]
          format: float
        as_of: { type: string, format: date-time }
        source: { type: string, examples: [polygon, coingecko] }
...
    ExportResource:
      type: string
      enum: [transactions, positions, snapshots, dividends]
    ExportCreateRequest:
      type: object
      required: [resource, format]
      additionalProperties: false
      properties:
        resource: { $ref: '#/components/schemas/ExportResource' }
        format:
          type: string
          enum: [csv]
        from:
          type: [string, "null"]
          format: date
        to:
          type: [string, "null"]
          format: date
        columns:
          type: array
          description: Optional column allow-list; if omitted, default set is exported.
          items: { type: string }
    ExportJob:
      type: object
      required: [id, resource, format, status, created_at]
      properties:
        id: { type: string, format: uuid }
        resource: { $ref: '#/components/schemas/ExportResource' }
        format: { type: string, enum: [csv] }
        status:
          type: string
          enum: [queued, running, done, failed]
        filters:
          type: object
          additionalProperties: true
        row_count:
          type: [integer, "null"]
          minimum: 0
        result_url:
          type: [string, "null"]
          format: uri
          description: Pre-signed URL; expires 24h after `completed_at`.
        error:
          type: [string, "null"]
        created_at: { type: string, format: date-time }
        completed_at:
          type: [string, "null"]
          format: date-time

    # ── Analytics (Pro) ───────────────────────────────────────
...